+++ package.json
@@ -14,7 +14,8 @@
     "lucide-react": "^0.292.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.18.0"
+    "react-router-dom": "^6.18.0",
+    "react-tooltip": "^5.22.0"
   },
   "devDependencies": {
     "@types/react": "^18.2.15",
