+++ src/App.jsx
@@ -2,9 +2,10 @@ import React, { useEffect, useState } from "react";
 import "./App.css";
 
 import { loader } from "./Utils/animations";
-import Preloader from "./Components/Preloader";
+import Preloader from "./Components/Preloader/Preloader";
 import { useNavigate } from "react-router-dom";
 import { Outlet } from "react-router-dom";
+
 
 function App() {
   const navigate = useNavigate();
