+++ src/Components/Button/Button.jsx
@@ -1,7 +1,10 @@
 
 import React from 'react'
 
-export default function Button({type, text, className, onClick}) {
+  
+
+export default function Button({  type, text, className, onClick}) {
+
   return (
     <button type={type} className={className} onClick={onClick}>{text}</button>
   )
