+++ src/pages/Dashboard/Dashboard.jsx
@@ -1,13 +1,14 @@
 import React from "react";
-import  Navbar  from "../../Components/Navbar";
+import  Navbar  from "../../Components/Navbar/Navbar";
 import { Outlet } from "react-router-dom";
+
 
 
 const Dashboard = () => {
   return (
     <>
      <div className="dashboard-page">
-      <Navbar />    
+      <Navbar />   
       <Outlet/>
     </div>
     </>
