+++ src/sass/_navbar.scss
@@ -30,6 +30,7 @@
     border: none;
     outline: none;
   }
+
   .home-active {
     width: 100%;
     height: 5vh;
